Replace DataLogger with DataStreamer2
