+++ src/ScottPlot5/ScottPlot5 Demos/ScottPlot5 WinForms Demo/Demos/DataLogger.cs
@@ -1,5 +1,4 @@
-﻿using ScottPlot;
-using ScottPlot.AxisPanels;
+﻿using ScottPlot.AxisPanels;
 
 namespace WinForms_Demo.Demos;
