+++ src/ScottPlot5/ScottPlot5/Plottables/DataLogger.cs
@@ -1,132 +1,159 @@
-﻿
-using ScottPlot.AxisLimitManagers;
+﻿using ScottPlot.AxisLimitManagers;
 using ScottPlot.DataSources;
 
 namespace ScottPlot.Plottables;
 
-public class DataLogger : IPlottable, IManagesAxisLimits, IHasLine, IHasLegendText
+public class DataLogger : IPlottable, IManagesAxisLimits, IHasLine, IHasMarker, IHasLegendText
 {
+    public DataLogger(IList<Coordinates> coordinates)
+    {
+        Data = new DataLoggerSource(coordinates);
+    }
+
+    public DataLogger()
+    {
+        Data = new DataLoggerSource(new List<Coordinates>());
+    }
+
+    public DataLoggerSource Data { get; }
+
     public bool IsVisible { get; set; } = true;
-    public IAxes Axes { get; set; } = ScottPlot.Axes.Default;
-    public DataLoggerSource Data { get; set; } = new();
-    public IEnumerable<LegendItem> LegendItems => LegendItem.Single(LegendText, LineStyle);
-
-    [Obsolete("use LegendText")]
-    public string Label { get => LegendText; set => LegendText = value; }
-    public string LegendText { get; set; } = string.Empty;
-
-    public bool ManageAxisLimits { get; set; } = true;
+    public IAxes Axes { get; set; } = new Axes();
     public IAxisLimitManager AxisManager { get; set; } = new Full();
-
-    public AxisLimits GetAxisLimits() => Data.GetAxisLimits();
-    public Color Color { get => LineStyle.Color; set => LineStyle.Color = value; }
-
-    /// <summary>
-    /// Returns true if data has been added since the last render
-    /// </summary>
-    public bool HasNewData => Data.CountTotal != Data.CountOnLastRender;
-
+    public bool Rotated { get; set; }
     public LineStyle LineStyle { get; set; } = new() { Width = 1 };
     public float LineWidth { get => LineStyle.Width; set => LineStyle.Width = value; }
     public LinePattern LinePattern { get => LineStyle.Pattern; set => LineStyle.Pattern = value; }
     public Color LineColor { get => LineStyle.Color; set => LineStyle.Color = value; }
 
-    public void UpdateAxisLimits(Plot plot)
+    public MarkerStyle MarkerStyle { get; set; } = new() { Size = 0, Shape = MarkerShape.FilledCircle };
+    public MarkerShape MarkerShape { get => MarkerStyle.Shape; set => MarkerStyle.Shape = value; }
+    public float MarkerSize { get => MarkerStyle.Size; set => MarkerStyle.Size = value; }
+    public Color MarkerFillColor { get => MarkerStyle.FillColor; set => MarkerStyle.FillColor = value; }
+    public Color MarkerLineColor { get => MarkerStyle.LineColor; set => MarkerStyle.LineColor = value; }
+    public Color MarkerColor { get => MarkerStyle.MarkerColor; set => MarkerStyle.MarkerColor = value; }
+    public float MarkerLineWidth { get => MarkerStyle.LineWidth; set => MarkerStyle.LineWidth = value; }
+
+    public bool HasNewData => Data.HasNewData;
+
+    /// <summary>
+    /// The style of lines to use when connecting points.
+    /// </summary>
+    public ConnectStyle ConnectStyle { get; set; } = ConnectStyle.Straight;
+
+    public Color Color
     {
-        if (Data.CountTotal == 0)
-            return;
-
-        bool firstTimeRenderingData = Data.CountOnLastRender < 1 && Data.CountTotal > 0;
-
-        AxisLimits dataLimits = GetAxisLimits();
-
-        AxisLimits viewLimits = firstTimeRenderingData
-            ? dataLimits
-            : plot.Axes.GetLimits(Axes.XAxis, Axes.YAxis);
-
-        AxisLimits newLimits = AxisManager.GetAxisLimits(viewLimits, dataLimits);
-
-        plot.Axes.SetLimits(newLimits, Axes.XAxis, Axes.YAxis);
+        get => LineStyle.Color;
+        set
+        {
+            LineStyle.Color = value;
+            MarkerStyle.FillColor = value;
+            MarkerStyle.LineColor = value;
+        }
     }
+
+    public double Period { get; set; } = 1.0;
+
+    [Obsolete("use LegendText")]
+    public string Label { get => LegendText; set => LegendText = value; }
+    public string LegendText { get; set; } = string.Empty;
+
+    public IEnumerable<LegendItem> LegendItems => LegendItem.Single(LegendText, LineStyle, MarkerStyle);
+
+    public AxisLimits GetAxisLimits()
+    {
+        CoordinateRange rangeX = Data.GetRangeX();
+        CoordinateRange rangeY = Data.GetRangeY(rangeX);
+
+        return Rotated
+            ? new AxisLimits(rangeY, rangeX)
+            : new AxisLimits(rangeX, rangeY);
+    }
+
+    public DataPoint GetNearest(Coordinates location, RenderDetails renderInfo, float maxDistance = 15) =>
+        Data.GetNearest(location, renderInfo, maxDistance);
+
+    public virtual void Render(RenderPack rp)
+    {
+        Pixel[] markerPixels = Rotated ? Data.GetPixelsToDrawVertically(rp, Axes, ConnectStyle)
+                                       : Data.GetPixelsToDrawHorizontally(rp, Axes, ConnectStyle);
+
+        Pixel[] linePixels = ConnectStyle switch
+        {
+            ConnectStyle.Straight => markerPixels,
+            ConnectStyle.StepHorizontal => Scatter.GetStepDisplayPixels(markerPixels, true),
+            ConnectStyle.StepVertical => Scatter.GetStepDisplayPixels(markerPixels, false),
+            _ => throw new NotSupportedException($"unsupported {nameof(ConnectStyle)}: {ConnectStyle}"),
+        };
+
+        using SKPaint paint = new();
+        Drawing.DrawLines(rp.Canvas, paint, linePixels, LineStyle);
+        Drawing.DrawMarkers(rp.Canvas, paint, markerPixels, MarkerStyle);
+
+        Data.OnRendered();
+    }
+
+    public void Add(Coordinates coordinates) => Data.Add(coordinates);
+
+    public void Add(double x, double y) => Add(new Coordinates(x, y));
 
     public void Add(double y)
     {
-        Data.Add(y);
-    }
-
-    public void Add(double x, double y)
-    {
-        Data.Add(x, y);
-    }
-
-    public void Add(double[] xs, double[] ys)
-    {
-        if (xs is null || ys is null)
-            throw new ArgumentException($"{nameof(xs)} and {nameof(ys)} must not be null");
-
-        if (xs.Length != ys.Length)
-            throw new ArgumentException($"{nameof(xs).Length} and {nameof(ys).Length} must have equal length");
-
-        for (int i = 0; i < xs.Length; i++)
-        {
-            Data.Add(xs[i], ys[i]);
-        }
-    }
-
-    public void Add(Coordinates coordinates)
-    {
-        Data.Add(coordinates);
-    }
-
-    public void Add(Coordinates[] coordinates)
-    {
-        if (coordinates is null)
-            throw new ArgumentException($"{coordinates} must not be null");
-
-        for (int i = 0; i < coordinates.Length; i++)
-        {
-            Data.Add(coordinates[i]);
-        }
+        double x = Data.Coordinates.Count == 0 ? 0 : Data.Coordinates[^1].X + Period;
+        Add(x, y);
     }
 
     public void Add(IEnumerable<double> ys)
     {
         foreach (double y in ys)
         {
-            Data.Add(y);
+            Add(y);
         }
     }
 
-    public void Add(IEnumerable<Coordinates> coordinates)
+    public void UpdateAxisLimits(Plot plot)
     {
-        foreach (Coordinates c in coordinates)
+        bool firstTimeRenderingData = !Data.WasRendered;
+
+        IAxis xAxis = Rotated ? Axes.YAxis : Axes.XAxis;
+        IAxis yAxis = Rotated ? Axes.XAxis : Axes.YAxis;
+
+        CoordinateRange dataRangeX = Data.GetRangeX();
+        CoordinateRange viewRangeX = firstTimeRenderingData
+            ? dataRangeX
+            : xAxis.GetRange().Rectified();
+
+        CoordinateRange newRangeX = AxisManager.GetRangeX(viewRangeX, dataRangeX);
+
+        // Get the Y range only for the newly calculated X range
+        CoordinateRange dataRangeY = Data.GetRangeY(newRangeX);
+        CoordinateRange viewRangeY = firstTimeRenderingData
+            ? dataRangeY
+            : yAxis.GetRange().Rectified();
+
+        CoordinateRange newRangeY = AxisManager.GetRangeY(viewRangeY, dataRangeY);
+
+        if (Rotated)
         {
-            Data.Add(c);
+            (newRangeX, newRangeY) = (newRangeY, newRangeX);
         }
+
+        if (Axes.XAxis.IsInverted())
+        {
+            newRangeX = new(newRangeX.Max, newRangeX.Min);
+        }
+
+        if (Axes.YAxis.IsInverted())
+        {
+            newRangeY = new(newRangeY.Max, newRangeY.Min);
+        }
+
+        AxisLimits newLimits = new(newRangeX, newRangeY);
+
+        plot.Axes.SetLimits(newLimits, Axes.XAxis, Axes.YAxis);
     }
 
-    public void Add(IReadOnlyList<double> xs, IReadOnlyList<double> ys)
-    {
-        if (xs.Count != ys.Count)
-        {
-            throw new ArgumentException($"{nameof(xs)} and {nameof(ys)} must have equal size");
-        }
-
-        for (int i = 0; i < xs.Count; i++)
-        {
-            Data.Add(xs[i], ys[i]);
-        }
-    }
-
-    public virtual void Render(RenderPack rp)
-    {
-        IEnumerable<Pixel> points = Data.Coordinates.Select(Axes.GetPixel);
-
-        using SKPaint paint = new();
-        Drawing.DrawLines(rp.Canvas, paint, points, LineStyle);
-
-        Data.CountOnLastRender = Data.CountTotal;
-    }
+    public bool ManageAxisLimits { get; set; } = true;
 
     /// <summary>
     /// Automatically expand the axis as needed to ensure the full dataset is visible before each render.
@@ -135,7 +162,7 @@ public class DataLogger : IPlottable, IManagesAxisLimits, IHasLine, IHasLegendTe
     {
         ManageAxisLimits = true;
         AxisManager = new Full();
-        Data.CountOnLastRender = -1;
+        Data.WasRendered = false;
     }
 
     /// <summary>
@@ -145,14 +172,8 @@ public class DataLogger : IPlottable, IManagesAxisLimits, IHasLine, IHasLegendTe
     public void ViewJump(double width = 1000, double paddingFraction = .5)
     {
         ManageAxisLimits = true;
-
-        AxisManager = new Slide()
-        {
-            Width = width,
-            PaddingFractionX = paddingFraction,
-        };
-
-        Data.CountOnLastRender = -1;
+        AxisManager = new Slide { Width = width, PaddingFractionX = paddingFraction };
+        Data.WasRendered = false;
     }
 
     /// <summary>
@@ -162,12 +183,7 @@ public class DataLogger : IPlottable, IManagesAxisLimits, IHasLine, IHasLegendTe
     public void ViewSlide(double width = 1000)
     {
         ManageAxisLimits = true;
-        AxisManager = new Slide()
-        {
-            Width = width,
-            PaddingFractionX = 0,
-        };
-        Data.CountOnLastRender = -1;
+        AxisManager = new Slide { Width = width, PaddingFractionX = 0 };
+        Data.WasRendered = false;
     }
-
 }
